Integrate autorotation into geolocation tracking controller
