+++ contribs/gmf/src/controllers/abstract.js
@@ -38,7 +38,6 @@ goog.require('ngeo.MapQuerent');
 goog.require('ngeo.StateManager');
 goog.require('ngeo.ToolActivate');
 goog.require('ngeo.ToolActivateMgr');
-goog.require('ol.DeviceOrientation');
 goog.require('ol.style.Circle');
 goog.require('ol.style.Fill');
 goog.require('ol.style.Stroke');
@@ -75,10 +74,6 @@ gmf.AbstractController = function(config, $scope, $injector) {
   if (this.ngeoLocation.hasParam('debug')) {
     // make the injector globally available
     window.injector = $injector;
-  }
-
-  if (config.autorotate) {
-    this.autorotateListener();
   }
 
   goog.asserts.assertInstanceof(this.map, ol.Map);
@@ -654,136 +649,5 @@ gmf.AbstractController.prototype.search_ = function(query, overlay) {
       }
     });
 };
-
-function isIOS() {
-  return /iPad|iPhone|iPod/.test(navigator.userAgent) &&
-    !window['MSStream'];
-}
-
-// Get heading depending on devices
-function headingFromDevices(deviceOrientation) {
-  let hdg = deviceOrientation.getHeading();
-  let orientation = window.orientation;
-  if (hdg === undefined) {
-    return undefined;
-  }
-  if (!isIOS()) {
-    hdg = -hdg;
-    if (window.screen.orientation.angle) {
-      orientation = window.screen.orientation.angle;
-    }
-  }
-  switch (orientation) {
-    case -90:
-      hdg = hdg - (Math.PI / 2);
-      break;
-
-    case 180:
-      hdg = hdg + Math.PI;
-      break;
-
-    case 90:
-      hdg = hdg + (Math.PI / 2);
-      break;
-
-    case 270:
-      hdg = hdg - (Math.PI / 2);
-      break;
-
-    default:
-      hdg = hdg;
-  }
-  return hdg;
-}
-
-// Update heading
-gmf.AbstractController.prototype.headingUpdate = function() {
-  let heading = headingFromDevices(this.deviceOrientation);
-  if (heading !== undefined) {
-    heading = -heading;
-    const currRotation = this.map.getView().getRotation();
-    const diff = heading - currRotation;
-
-    if (diff > Math.PI) {
-      heading -= 2 * Math.PI;
-    }
-    this.map.getView().animate({
-      rotation: heading,
-      duration: 350,
-      easing: ol.easing.linear
-    });
-  }
-};
-
-// Orientation control events
-gmf.AbstractController.prototype.autorotateListener = function() {
-  this.deviceOrientation = new ol.DeviceOrientation();
-
-  console.log(1, this.deviceOrientation);
-
-  let currHeading = 0;
-  const headngUpdateWhenMapRotate = throttle(this.headingUpdate, 300, this);
-
-  this.deviceOrientation.on(['change'], (event) => {
-    const heading = headingFromDevices(this.deviceOrientation);
-    if (heading === undefined) {
-      console.error('Heading is undefined');
-      return;
-    }
-
-    if (Math.abs(heading - currHeading) > 0.05) {
-      currHeading = heading;
-      headngUpdateWhenMapRotate();
-    }
-  });
-
-  this.deviceOrientation.setTracking(true);
-};
-
-function throttle(fn, time, context) {
-  let lock, args, asyncKey, destroyed;
-
-  function later() {
-    // reset lock and call if queued
-    lock = false;
-    if (args) {
-      throttled.call(context, args);
-      args = false;
-    }
-  }
-
-  const checkDestroyed = function() {
-    if (destroyed) {
-      throw new Error('Method was already destroyed');
-    }
-  };
-
-  function throttled(...argumentList) {
-    checkDestroyed();
-
-    if (lock) {
-      // called too soon, queue to call later
-      args = argumentList;
-      return;
-    }
-
-    // call and lock until later
-    fn.apply(context, argumentList);
-    asyncKey = setTimeout(later, time);
-    lock = true;
-  }
-
-  throttled.destroy = function() {
-    checkDestroyed();
-
-    if (asyncKey) {
-      clearTimeout(asyncKey);
-    }
-
-    destroyed = true;
-  };
-
-  return throttled;
-}
 
 gmf.module.controller('AbstractController', gmf.AbstractController);
