+++ src/directives/mobilegeolocation.js
@@ -11,6 +11,7 @@ goog.require('ol.GeolocationProperty');
 goog.require('ol.Map');
 goog.require('ol.ViewProperty');
 goog.require('ol.geom.Point');
+goog.require('ol.DeviceOrientation');
 
 
 /**
@@ -120,10 +121,23 @@ ngeo.MobileGeolocationController = function($scope, $element,
       enableHighAccuracy: true
     })
   });
+  
+  /**
+   * @private
+   * @type {ol.DeviceOrientation}
+   */
+  this.deviceOrientation;
+
+  if (options.autorotate) {
+    this.autorotateListener();
+  }
 
   // handle geolocation error.
   this.geolocation_.on('error', function(error) {
     this.untrack_();
+    if (this.deviceOrientation) {
+      this.deviceOrientation.setTracking(false);
+    }
     let msg;
     switch (error.code) {
       case 1:
@@ -216,11 +230,11 @@ ngeo.MobileGeolocationController = function($scope, $element,
     this.handleViewChange_,
     this);
 
-  ol.events.listen(
-    view,
-    ol.Object.getChangeEventType(ol.ViewProperty.ROTATION),
-    this.handleViewChange_,
-    this);
+  // ol.events.listen(
+  //   view,
+  //   ol.Object.getChangeEventType(ol.ViewProperty.ROTATION),
+  //   this.handleViewChange_,
+  //   this);
 
   ngeoDecorateGeolocation(this.geolocation_);
 };
@@ -311,6 +325,137 @@ ngeo.MobileGeolocationController.prototype.handleViewChange_ = function(event) {
     this.follow_ = false;
   }
 };
+
+function isIOS() {
+  return /iPad|iPhone|iPod/.test(navigator.userAgent) &&
+    !window['MSStream'];
+}
+
+// Get heading depending on devices
+function headingFromDevices(deviceOrientation) {
+  let hdg = deviceOrientation.getHeading();
+  let orientation = window.orientation;
+  if (hdg === undefined) {
+    return undefined;
+  }
+  if (!isIOS()) {
+    hdg = -hdg;
+    if (window.screen.orientation.angle) {
+      orientation = window.screen.orientation.angle;
+    }
+  }
+  switch (orientation) {
+    case -90:
+      hdg = hdg - (Math.PI / 2);
+      break;
+
+    case 180:
+      hdg = hdg + Math.PI;
+      break;
+
+    case 90:
+      hdg = hdg + (Math.PI / 2);
+      break;
+
+    case 270:
+      hdg = hdg - (Math.PI / 2);
+      break;
+
+    default:
+      hdg = hdg;
+  }
+  return hdg;
+}
+
+// Update heading
+ngeo.MobileGeolocationController.prototype.headingUpdate = function() {
+  let heading = headingFromDevices(this.deviceOrientation);
+  if (heading !== undefined) {
+    heading = -heading;
+    const currRotation = this.map_.getView().getRotation();
+    const diff = heading - currRotation;
+
+    if (diff > Math.PI) {
+      heading -= 2 * Math.PI;
+    }
+    this.map_.getView().animate({
+      rotation: heading,
+      duration: 350,
+      easing: ol.easing.linear
+    });
+  }
+};
+
+// Orientation control events
+ngeo.MobileGeolocationController.prototype.autorotateListener = function() {
+  this.deviceOrientation = new ol.DeviceOrientation();
+
+  console.log(1, this.deviceOrientation);
+
+  let currHeading = 0;
+  const headngUpdateWhenMapRotate = throttle(this.headingUpdate, 300, this);
+
+  this.deviceOrientation.on(['change'], (event) => {
+    const heading = headingFromDevices(this.deviceOrientation);
+    if (heading === undefined) {
+      console.error('Heading is undefined');
+      return;
+    }
+
+    if (Math.abs(heading - currHeading) > 0.05) {
+      currHeading = heading;
+      headngUpdateWhenMapRotate();
+    }
+  });
+
+  this.deviceOrientation.setTracking(true);
+};
+
+function throttle(fn, time, context) {
+  let lock, args, asyncKey, destroyed;
+
+  function later() {
+    // reset lock and call if queued
+    lock = false;
+    if (args) {
+      throttled.call(context, args);
+      args = false;
+    }
+  }
+
+  const checkDestroyed = function() {
+    if (destroyed) {
+      throw new Error('Method was already destroyed');
+    }
+  };
+
+  function throttled(...argumentList) {
+    checkDestroyed();
+
+    if (lock) {
+      // called too soon, queue to call later
+      args = argumentList;
+      return;
+    }
+
+    // call and lock until later
+    fn.apply(context, argumentList);
+    asyncKey = setTimeout(later, time);
+    lock = true;
+  }
+
+  throttled.destroy = function() {
+    checkDestroyed();
+
+    if (asyncKey) {
+      clearTimeout(asyncKey);
+    }
+
+    destroyed = true;
+  };
+
+  return throttled;
+}
 
 
 ngeo.module.controller('NgeoMobileGeolocationController', ngeo.MobileGeolocationController);
